# public/swagger.json
{
	"components": {
...
		"schemas": {
			"SaveUpdateResItem": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"item_name": {
						"type": "string"
					},
					"item_price": {
						"type": "string"
					},
					"item_size": {
						"type": "string"
					},
					"item_type": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"item_name",
					"item_price",
					"item_size",
					"item_type",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetItem": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getItemPrice": {
				"properties": {
					"item_price": {
						"type": "string"
					}
				},
				"required": [
					"item_price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqItem": {
				"properties": {
					"item_name": {
						"type": "string"
					},
					"item_price": {
						"type": "string"
					},
					"item_size": {
						"type": "string"
					},
					"item_type": {
						"type": "string"
					}
				},
				"required": [
					"item_name",
					"item_price",
					"item_size",
					"item_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqItem": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"item_name": {
						"type": "string"
					},
					"item_price": {
						"type": "string"
					},
					"item_size": {
						"type": "string"
					},
					"item_type": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"item_name",
					"item_price",
					"item_size",
					"item_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteItem": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResOrder": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"tableNo": {
						"type": "string"
					},
					"item": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"waiter": {
						"type": "string"
					},
					"o_status": {
						"type": "string"
					},
					"o_discription": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"tableNo",
					"item",
					"waiter",
					"o_status",
					"o_discription",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SearchReqOrder": {
				"properties": {
					"item": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"item"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetOrderByStatus": {
				"properties": {
					"o_status": {
						"type": "string"
					}
				},
				"required": [
					"o_status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResWaiter": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"waiter_name": {
						"type": "string"
					},
					"waiter_email": {
						"type": "string"
					},
					"waiter_password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"waiter_name",
					"waiter_email",
					"waiter_password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SaveReqWaiter": {
				"properties": {
					"waiter_name": {
						"type": "string"
					},
					"waiter_email": {
						"type": "string"
					},
					"waiter_password": {
						"type": "string"
					}
				},
				"required": [
					"waiter_name",
					"waiter_email",
					"waiter_password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"admin_name": {
						"type": "string"
					},
					"admin_email": {
						"type": "string"
					},
					"admin_password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"admin_name",
					"admin_email",
					"admin_password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqAdmin": {
				"properties": {
					"admin_name": {
						"type": "string"
					},
					"admin_email": {
						"type": "string"
					},
					"admin_password": {
						"type": "string"
					}
				},
				"required": [
					"admin_name",
					"admin_email",
					"admin_password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAdmin": {
				"properties": {
					"admin_email": {
						"type": "string"
					},
					"admin_password": {
						"type": "string"
					}
				},
				"required": [
					"admin_email",
					"admin_password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "http",
				"name": "Authorization",
				"in": "header",
				"scheme": "Bearer"
			}
		}
	},
	"info": {
		"title": "hotel-management-system",
		"version": "1.0.0",
		"description": "Hotel Management System by - Muhammad Ahsan",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Items/getItemByID": {
			"post": {
				"operationId": "GetItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResItem"
								}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetItem"
							}
						}
					}
				}
			}
		},
		"/Items/getMenu": {
			"post": {
				"operationId": "GetItemList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get list of all the items => anyone is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Items/getItemsbyPrice": {
			"post": {
				"operationId": "GetItemByPrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get list of all the items with spacific price=> anyone is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/getItemPrice"
							}
						}
					}
				}
			}
		},
		"/Items/saveItemInMenu": {
			"post": {
				"operationId": "SaveItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResItem"
								}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqItem"
							}
						}
					}
				}
			}
		},
		"/Items/updateItemInMenu": {
			"put": {
				"operationId": "UpdateItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResItem"
								}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqItem"
							}
						}
					}
				}
			}
		},
		"/Items/deleteItem": {
			"delete": {
				"operationId": "DeletItem",
				"responses": {
					"200": {
						"description": "product deleted",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteItem"
							}
						}
					}
				}
			}
		},
...
		"/orders/getorderlist": {
			"post": {
				"operationId": "GetorderList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get list of all the orders => Only Login admin is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/orders/getorderbyItemID": {
			"post": {
				"operationId": "SearchorderByproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get all the orders contain certain item => Only Login admin is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchReqOrder"
							}
						}
					}
				}
			}
		},
		"/orders/getOrderByStatus": {
			"post": {
				"operationId": "GetorderByStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrderByStatus"
							}
						}
					}
				}
			}
		},
...
		"/waiter/Registerwaiter": {
			"post": {
				"operationId": "SaveWaiter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResWaiter"
								}
							}
						}
					}
				},
				"summary": "Register new waiter => Only Login admin is allowed",
				"tags": [
					"PMS - Waiter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqWaiter"
							}
						}
					}
				}
			}
		},
		"/admin/RegisterAdmin": {
			"post": {
				"operationId": "SaveAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"summary": "Register as an Admin",
				"tags": [
					"PMS - Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/adminlogin": {
			"post": {
				"operationId": "AuthAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"summary": "Login as an Admin",
				"tags": [
					"PMS - Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetAdmin"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}