# src/types/Request/waiter.request.ts
export interface SaveReqWaiter {
  waiter_name: string,
  waiter_email:string, 
  waiter_password: string,
  }


export interface GetWaiter {
    waiter_email:string, 
    waiter_password:string, 
  }
  // export interface UpdateReqProduct {
  //   _id: string;
  //   p_name: String,
  //   p_price:String, 
  //   p_discription: String,
  // }
  // export interface Getproduct {
  //   id: string
  // }
  // export interface search {
  //   p_name: string
  // }
  // export interface DeleteProduct {
  //   id: string
  // }
  // export interface getProductPrice{
 
  //   p_price:String, 
    
  //   }
 